% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDEfunctions.R
\name{coef.IDE}
\alias{coef.IDE}
\title{Retrieve estimated regression coefficients}
\usage{
\method{coef}{IDE}(object, ...)
}
\arguments{
\item{object}{object of class \code{IDE}}

\item{...}{currently unused}
}
\description{
Takes a an object of class \code{IDE} and returns a numeric vector with the estimated regression coefficients.
}
\examples{
SIM1 <- simIDE(T = 5, nobs = 100, k_spat_invariant = 1)
coef(SIM1$IDEmodel)
}
\seealso{
\code{\link{IDE}} for more information on how to construct and fit an IDE model.
}
