% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDEfunctions.R
\name{constant_basis}
\alias{constant_basis}
\title{Create a single, constant basis function}
\usage{
constant_basis()
}
\value{
Object of class \code{Basis}
}
\description{
Constructs an object of class \code{Basis} as defined in \code{FRK} that is constant over the entire spatial domain.
}
\examples{
basis1 <- constant_basis()
}
\seealso{
\code{\link{IDE}} for how to use basis functions to construct the IDE kernel
}
