% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDEfunctions.R
\name{symb}
\alias{symb}
\title{Return the symbolic representation of a Matrix}
\usage{
symb(A)
}
\arguments{
\item{A}{object of class Matrix}
}
\value{
object of class Matrix
}
\description{
This function takes an object of class Matrix and returns the same Matrix with all elements replaced with 1
#'
}
\examples{
require(Matrix)
Q <- sparseMatrix(i=c(1,2,2),j=c(1,1,2),x=c(0.1,0.2,1))
Qsymb <- symb(Q)
Qsymb
}
